requirements.txt
ffmpeg-python<=0.2.0
numpy>=1.11.0
pandas==1.3.5
opencv-python-headless<=4.2.0.32
ffmpeg<=4.2.2
tensorflow-cpu==2.6.*
Flask<=2.1.3
protobuf==3.20.*
gunicorn